Remove dash support
Dashes, padding, and trailing the check character have surprising
combinations which makes the output of an encode call harder to predict.

For example: with all three the dash interval may not be respected for
part of the encoded value. This would make writing splitting code much
more difficult for a user.